+++ packages/browser-wallet/src/popup/page-layouts/MainLayout/i18n/en.ts
@@ -16,6 +16,9 @@ const t = {
     accountList: {
         new: 'Add new',
     },
+    identityList: {
+        new: 'Request new',
+    },
 };
 
 export default t;