+++ packages/browser-wallet/src/popup/store/settings.ts
@@ -1,8 +1,12 @@
-import { ChromeStorageKey, EncryptedData, Theme, WalletCredential } from '@shared/storage/types';
+import { Buffer } from 'buffer/';
+import { ChromeStorageKey, EncryptedData, NetworkConfiguration, Theme, WalletCredential } from '@shared/storage/types';
 import { atom } from 'jotai';
 import { EventType } from '@concordium/browser-wallet-api-helpers';
 import { popupMessageHandler } from '@popup/shared/message-handler';
-import { atomWithChromeStorage, AsyncWrapper } from './utils';
+import { decrypt } from '@popup/shared/crypto';
+import { mnemonicToSeedSync } from '@scure/bip39';
+import { mainnet } from '@popup/pages/NetworkSettings/NetworkSettings';
+import { atomWithChromeStorage } from './utils';
 
 export const credentialsAtom = atomWithChromeStorage<WalletCredential[]>(ChromeStorageKey.Credentials, []);
 export const encryptedSeedPhraseAtom = atomWithChromeStorage<EncryptedData | undefined>(
@@ -12,18 +16,15 @@ export const encryptedSeedPhraseAtom = atomWithChromeStorage<EncryptedData | und
 );
 export const themeAtom = atomWithChromeStorage<Theme>(ChromeStorageKey.Theme, Theme.Light);
 
-const storedJsonRpcUrlAtom = atomWithChromeStorage<string>(ChromeStorageKey.JsonRpcUrl, '', true);
-export const jsonRpcUrlAtomLoading = atom<AsyncWrapper<string>, string>(
-    (get) => get(storedJsonRpcUrlAtom),
-    (_, set, jsonRpcUrl) => {
-        set(storedJsonRpcUrlAtom, jsonRpcUrl);
-        popupMessageHandler.broadcast(EventType.ChainChanged, jsonRpcUrl);
-    }
+const storedNetworkConfigurationAtom = atomWithChromeStorage<NetworkConfiguration>(
+    ChromeStorageKey.NetworkConfiguration,
+    mainnet
 );
-export const jsonRpcUrlAtom = atom<string, string>(
-    (get) => get(jsonRpcUrlAtomLoading).value,
-    (_, set, jsonRpcUrl) => {
-        set(jsonRpcUrlAtomLoading, jsonRpcUrl);
+export const networkConfigurationAtom = atom<NetworkConfiguration, NetworkConfiguration>(
+    (get) => get(storedNetworkConfigurationAtom),
+    (_, set, networkConfiguration) => {
+        set(storedNetworkConfigurationAtom, networkConfiguration);
+        popupMessageHandler.broadcast(EventType.ChainChanged, networkConfiguration.genesisHash);
     }
 );
 
@@ -31,4 +32,19 @@ export const sessionPasscodeAtom = atomWithChromeStorage<string | undefined>(
     ChromeStorageKey.Passcode,
     undefined,
     true
+);
+
+export const seedPhraseAtom = atom<string, never>(
+    (get) => {
+        const seed = get(encryptedSeedPhraseAtom).value;
+        const passcode = get(sessionPasscodeAtom).value;
+
+        if (seed && passcode) {
+            return Buffer.from(mnemonicToSeedSync(decrypt(seed, passcode))).toString('hex');
+        }
+        throw new Error('SeedPhrase should not be retrieved without unlocking the wallet.');
+    },
+    () => {
+        throw new Error('Setting the seedPhrase directly is not supported');
+    }
 );