+++ packages/browser-wallet/src/popup/store/utils.ts
@@ -3,21 +3,31 @@ import {
     StorageAccessor,
     storedConnectedSites,
     storedCredentials,
+    storedCurrentNetwork,
     storedEncryptedSeedPhrase,
-    storedJsonRpcUrl,
     storedSelectedAccount,
     storedTheme,
+    storedIdentities,
+    storedPendingIdentity,
+    storedSelectedIdentity,
+    storedSeedPhrase,
+    storedIdentityProviders,
 } from '@shared/storage/access';
 import { ChromeStorageKey } from '@shared/storage/types';
 import { atom, WritableAtom } from 'jotai';
 
 const accessorMap = {
+    [ChromeStorageKey.Identities]: storedIdentities,
+    [ChromeStorageKey.PendingIdentity]: storedPendingIdentity,
+    [ChromeStorageKey.SelectedIdentity]: storedSelectedIdentity,
     [ChromeStorageKey.ConnectedSites]: storedConnectedSites,
     [ChromeStorageKey.Credentials]: storedCredentials,
     [ChromeStorageKey.SelectedAccount]: storedSelectedAccount,
-    [ChromeStorageKey.JsonRpcUrl]: storedJsonRpcUrl,
     [ChromeStorageKey.SeedPhrase]: storedEncryptedSeedPhrase,
+    [ChromeStorageKey.NetworkConfiguration]: storedCurrentNetwork,
     [ChromeStorageKey.Theme]: storedTheme,
+    [ChromeStorageKey.SeedPhrase]: storedSeedPhrase,
+    [ChromeStorageKey.IdentityProviders]: storedIdentityProviders,
     [ChromeStorageKey.Passcode]: sessionPasscode,
 };